+++ doc/release/RELEASE-NOTES.md
@@ -17,11 +17,12 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 * [OSDEV-1229](https://opensupplyhub.atlassian.net/browse/OSDEV-1229) - Created Moderation Events Postgres table to track moderation events in the database.
 
 ### Code/API changes
-* Throttling has been introduced for tiles/* endpoints, limiting requests to 300 per minute.
+* Throttling has been introduced for tiles/* endpoints, limiting requests to 30 per minute.
 * [OSDEV-1328](https://opensupplyhub.atlassian.net/browse/OSDEV-1328) The OpenSearch tokenizer has been changed to `lowercase` to get better search results when querying the GET /v1/production-locations/ endpoint.
 
 ### Architecture/Environment changes
 * Resource allocation has been optimized for the staging environment. The number of ECS tasks for the Django app has been reduced from 6 to 4, while maintaining system stability.
+
 
 ### Release instructions:
 * Ensure that the following commands are included in the `post_deployment` command: